requirements.txt
tensorflow-gpu==2.9.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
keras==2.9.0
